updated form class to function
